pyproject.toml
[tool.poetry]
name = "techyon-login"
version = "0.1.0"
description = ""
authors = ["sohamjani <soham.sanjay@otocapital.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
ipython = "^8.24.0"
graphviz = "^0.20.3"
django-extensions = "^3.2.3"
pyparsing = "^3.1.2"
pydot = "^2.0.0"
django-environ = "^0.11.2"
django-choices = "^2.0.0"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
